// client/packages/annotator/src/annotator.types.tsx
export type Maybe<T> = T | null | undefined;
// ...
export interface ToolSettingsBBOX {
    color: string;
}
// ...
export interface MouseEvent {
    point: paper.Point;
    item: paper.Item;
    modifiers?: {
        shift?: boolean;
    };
}